backend/requirements.txt
fastapi
uvicorn
google-cloud-speech
google-cloud-texttospeech
python-multipart
databases
aiosqlite
openai
python-dotenv